+++ @utahdts/utah-design-system/react/components/table/TableFilterDate.jsx
@@ -10,6 +10,7 @@ import { useCurrentValuesFromStateContext } from './useCurrentValuesFromStateCon
  * @param {TableDataT} [props.defaultValue]
  * @param {import('react').RefObject<HTMLTableCellElement>} [props.innerRef]
  * @param {string} [props.id]
+ * @param {string} props.label
  * @param {(e: React.ChangeEvent) => TableDataT} [props.onChange]
  * @param {string} props.recordFieldPath
  * @param {TableDataT} [props.value]
@@ -20,6 +21,7 @@ export function TableFilterDate({
   defaultValue,
   innerRef,
   id,
+  label,
   onChange,
   recordFieldPath,
   value,
@@ -40,7 +42,7 @@ export function TableFilterDate({
     <th className={joinClassNames('table-header__cell table-header__cell--filter-date', className)} id={id ?? undefined} ref={innerRef}>
       <TextInput
         id={`table-filter-date-${recordFieldPath}`}
-        label={`filter ${recordFieldPath}`}
+        label={label}
         onChange={currentOnChange}
         value={currentValue?.toString()}
         // eslint-disable-next-line react/jsx-props-no-spreading